if we collect not in level

+++ src/com/watabou/pixeldungeon/actors/hero/Hero.java
@@ -1450,7 +1450,9 @@ public class Hero extends Char {
 	
 	public void collect(Item item) {
 		if(!item.collect(this)){
-			Dungeon.level.drop(item, pos).sprite.drop();
+			if(Dungeon.level!=null){
+				Dungeon.level.drop(item, pos).sprite.drop();
+			}
 		}
 	}
 }
